Refactor MAT filters table to virtualize rows
