+++ package.json
@@ -47,7 +47,6 @@
     "react-sticky": "^6.0.3",
     "react-table": "^7.8.0",
     "react-virtual": "^2.10.4",
-    "react-virtualized-auto-sizer": "^1.0.6",
     "react-window": "^1.8.7",
     "rebass": "4.0.7",
     "styled-components": "5.1.1",
